refactor and correct paths
