+++ src/containers/Interfaces/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
 export { default as NameGeneratorInterface } from './NameGenerator';
-export { default as NameGeneratorListInterface } from './NameGeneratorListInterface';
+export { default as NameGeneratorListInterface } from './NameGeneratorList';
 export { default as SociogramInterface } from './Sociogram';
 export { default as QuizInterface } from './Quiz';
